refactor: remove compat re-exports from cella-config lib.rs
Remove `pub use devcontainer::*` and `pub use settings::{...}` from
crate root. Fix all internal `crate::` paths to `super::` within
devcontainer module. Update consumers to use qualified paths.
