+++ crates/cella-cli/src/commands/features/update.rs
@@ -51,7 +51,7 @@ impl UpdateArgs {
     pub async fn execute(self) -> Result<(), Box<dyn std::error::Error>> {
         let config_path = resolve::discover_config(&self.common)?;
         let raw = resolve::read_raw_config(&config_path)?;
-        let stripped = cella_config::jsonc::strip(&raw)?;
+        let stripped = cella_config::devcontainer::jsonc::strip(&raw)?;
         let config: serde_json::Value = serde_json::from_str(&stripped)?;
         let features = resolve::extract_features(&config);
